backend/tests/requirements-test.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
coverage==7.3.2
Flask-Testing==0.8.1 